+++ common-rest/src/main/java/org/uniprot/api/common/repository/search/SolrQueryRepository.java
@@ -1,5 +1,10 @@
 package org.uniprot.api.common.repository.search;
 
+import java.io.IOException;
+import java.util.*;
+import java.util.stream.Stream;
+import java.util.stream.StreamSupport;
+
 import org.apache.solr.client.solrj.SolrClient;
 import org.apache.solr.client.solrj.SolrQuery;
 import org.apache.solr.client.solrj.SolrServerException;
@@ -20,11 +25,6 @@ import org.uniprot.api.common.repository.search.term.TermInfo;
 import org.uniprot.api.common.repository.search.term.TermInfoConverter;
 import org.uniprot.store.search.SolrCollection;
 import org.uniprot.store.search.document.Document;
-
-import java.io.IOException;
-import java.util.*;
-import java.util.stream.Stream;
-import java.util.stream.StreamSupport;
 
 /**
  * Solr Basic Repository class to enable the execution of dynamically build queries in a solr
@@ -125,8 +125,9 @@ public abstract class SolrQueryRepository<T extends Document> {
     public QueryResponse query(SolrQuery solrQuery) {
         try {
             return solrClient.query(collection.name(), solrQuery);
-        } catch (Exception e) {
-            throw new QueryRetrievalException("Unexpected error retrieving data from our Repository", e);
+        } catch (SolrServerException | IOException e) {
+            throw new QueryRetrievalException(
+                    "Unexpected error retrieving data from our Repository", e);
         }
     }
