+++ common-rest/src/main/java/org/uniprot/api/rest/request/keyword/KeywordBasicRequest.java
@@ -1,7 +1,9 @@
 package org.uniprot.api.rest.request.keyword;
 
-import io.swagger.v3.oas.annotations.Parameter;
+import javax.validation.constraints.NotNull;
+
 import lombok.Data;
+
 import org.uniprot.api.rest.request.QueryFieldMetaReaderImpl;
 import org.uniprot.api.rest.request.ReturnFieldMetaReaderImpl;
 import org.uniprot.api.rest.request.SortFieldMetaReaderImpl;
@@ -10,9 +12,9 @@ import org.uniprot.api.rest.validation.ValidSolrQueryFields;
 import org.uniprot.api.rest.validation.ValidSolrQuerySyntax;
 import org.uniprot.api.rest.validation.ValidSolrSortFields;
 import org.uniprot.store.config.UniProtDataType;
-import uk.ac.ebi.uniprot.openapi.extension.ModelFieldMeta;
 
-import javax.validation.constraints.NotNull;
+import uk.ac.ebi.uniprot.openapi.extension.ModelFieldMeta;
+import io.swagger.v3.oas.annotations.Parameter;
 
 /**
  * @author sahmad
