+++ common-rest/src/main/java/org/uniprot/api/rest/request/taxonomy/GetByTaxonIdsRequest.java
@@ -1,22 +1,25 @@
 package org.uniprot.api.rest.request.taxonomy;
 
-import io.swagger.v3.oas.annotations.Parameter;
+import java.util.Arrays;
+import java.util.List;
+import java.util.stream.Collectors;
+
+import javax.validation.constraints.Max;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+import javax.validation.constraints.PositiveOrZero;
+
 import lombok.Data;
+
 import org.springframework.http.MediaType;
 import org.uniprot.api.rest.request.ReturnFieldMetaReaderImpl;
 import org.uniprot.api.rest.request.SearchRequest;
 import org.uniprot.api.rest.validation.*;
 import org.uniprot.core.util.Utils;
 import org.uniprot.store.config.UniProtDataType;
-import uk.ac.ebi.uniprot.openapi.extension.ModelFieldMeta;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.PositiveOrZero;
-import java.util.Arrays;
-import java.util.List;
-import java.util.stream.Collectors;
+import uk.ac.ebi.uniprot.openapi.extension.ModelFieldMeta;
+import io.swagger.v3.oas.annotations.Parameter;
 
 /**
  * @author lgonzales
