+++ common-rest/src/main/java/org/uniprot/api/rest/request/taxonomy/TaxonomyFacetConfig.java
@@ -1,16 +1,17 @@
 package org.uniprot.api.rest.request.taxonomy;
 
+import java.util.Collection;
+import java.util.HashMap;
+import java.util.Map;
+
 import lombok.Getter;
 import lombok.Setter;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Component;
 import org.uniprot.api.common.repository.search.facet.FacetConfig;
 import org.uniprot.api.common.repository.search.facet.FacetProperty;
-
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
 
 @Component
 @Getter
