+++ common-rest/src/main/java/org/uniprot/api/rest/request/taxonomy/TaxonomySolrQueryConfig.java
@@ -1,5 +1,7 @@
 package org.uniprot.api.rest.request.taxonomy;
 
+import java.util.*;
+
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.uniprot.api.common.repository.search.SolrQueryConfig;
@@ -11,8 +13,6 @@ import org.uniprot.store.config.UniProtDataType;
 import org.uniprot.store.config.searchfield.common.SearchFieldConfig;
 import org.uniprot.store.config.searchfield.factory.SearchFieldConfigFactory;
 import org.uniprot.store.config.searchfield.model.SearchFieldItem;
-
-import java.util.*;
 
 @Configuration
 public class TaxonomySolrQueryConfig {
