+++ common-rest/src/main/java/org/uniprot/api/rest/request/taxonomy/TaxonomySortClause.java
@@ -1,11 +1,11 @@
 package org.uniprot.api.rest.request.taxonomy;
 
+import javax.annotation.PostConstruct;
+
 import org.apache.solr.client.solrj.SolrQuery;
 import org.springframework.stereotype.Component;
 import org.uniprot.api.rest.search.AbstractSolrSortClause;
 import org.uniprot.store.config.UniProtDataType;
-
-import javax.annotation.PostConstruct;
 
 @Component
 public class TaxonomySortClause extends AbstractSolrSortClause {
