+++ common-rest/src/main/java/org/uniprot/api/rest/request/taxonomy/TaxonomyStreamRequest.java
@@ -1,11 +1,13 @@
 package org.uniprot.api.rest.request.taxonomy;
 
-import io.swagger.v3.oas.annotations.Parameter;
+import javax.validation.constraints.Pattern;
+
 import lombok.Data;
 import lombok.EqualsAndHashCode;
+
 import org.uniprot.api.rest.request.StreamRequest;
 
-import javax.validation.constraints.Pattern;
+import io.swagger.v3.oas.annotations.Parameter;
 
 /**
  * @author sahmad
