+++ common-rest/src/main/java/org/uniprot/api/rest/response/taxonomy/TaxonomyStreamConfig.java
@@ -6,8 +6,8 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 import org.uniprot.api.common.repository.stream.document.DefaultDocumentIdStream;
-import org.uniprot.api.rest.respository.RepositoryConfig;
 import org.uniprot.api.rest.converter.SolrDocumentRdfIdConverter;
+import org.uniprot.api.rest.respository.RepositoryConfig;
 import org.uniprot.api.rest.respository.taxonomy.TaxonomyRepository;
 import org.uniprot.store.search.document.taxonomy.TaxonomyDocument;
