+++ common-rest/src/main/java/org/uniprot/api/rest/service/BasicSearchService.java
@@ -1,9 +1,17 @@
 package org.uniprot.api.rest.service;
 
+import static org.uniprot.api.rest.output.PredefinedAPIStatus.LEADING_WILDCARD_IGNORED;
+
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.Set;
+import java.util.function.Function;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
 import org.apache.commons.lang3.math.NumberUtils;
-import org.apache.solr.client.solrj.SolrQuery;
-import org.apache.solr.client.solrj.response.FacetField;
-import org.apache.solr.common.params.FacetParams;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.PropertySource;
 import org.uniprot.api.common.exception.ResourceNotFoundException;
@@ -23,14 +31,6 @@ import org.uniprot.store.config.searchfield.model.SearchFieldItem;
 import org.uniprot.store.search.SolrQueryUtil;
 import org.uniprot.store.search.document.Document;
 import org.uniprot.store.search.field.validator.FieldRegexConstants;
-
-import java.util.*;
-import java.util.function.Function;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
-
-import static org.uniprot.api.rest.output.PredefinedAPIStatus.LEADING_WILDCARD_IGNORED;
 
 /**
  * @param <D> the type of the input to the class. a type of Document
@@ -218,7 +218,7 @@ public abstract class BasicSearchService<D extends Document, R> {
         return requestBuilder;
     }
 
-    private String getQueryFields(String query) {
+    protected String getQueryFields(String query) {
         String queryFields = "";
         Optional<String> optimisedQueryField = validateOptimisableField(query);
         if (optimisedQueryField.isPresent()) {
@@ -258,15 +258,6 @@ public abstract class BasicSearchService<D extends Document, R> {
     public String getRdf(String id, String dataType, String format) {
         return getRdfStreamer().stream(Stream.of(id), dataType, format)
                 .collect(Collectors.joining());
-    }
-
-    public List<FacetField> getFacets(String query, Map<String,String> facetFields) {
-        SolrQuery solrQuery = new SolrQuery(query);
-        facetFields.forEach(solrQuery::set);
-        solrQuery.set(FacetParams.FACET, true);
-        solrQuery.set(DEF_TYPE, "edismax");
-        solrQuery.add(QUERY_FIELDS, getQueryFields(query));
-        return repository.query(solrQuery).getFacetFields();
     }
 
     protected RdfStreamer getRdfStreamer() {
