+++ uniprotkb-rest/src/main/java/org/uniprot/api/uniprotkb/view/ViewBy.java
@@ -3,11 +3,11 @@ package org.uniprot.api.uniprotkb.view;
 import java.util.Comparator;
 
 public interface ViewBy {
-    public String getId();
-    public String getLabel();
-    public String getLink();
-    public boolean isExpand();
-    public long getCount();
+    String getId();
+    String getLabel();
+    String getLink();
+    boolean isExpand();
+    long getCount();
 
     Comparator<ViewBy> SORT_BY_LABEL_IGNORE_CASE = Comparator.comparing(ViewBy::getLabel, String::compareToIgnoreCase);
     Comparator<ViewBy> SORT_BY_ID = Comparator.comparing(ViewBy::getId);
